+++ src/modules/audio/null/Source.h
@@ -77,7 +77,7 @@ public:
 	virtual int getChannels() const;
 
 	virtual int getFreeBufferCount() const;
-	virtual bool queue(void *data, int length, int dataSampleRate, int dataBitDepth, int dataChannels);
+	virtual bool queue(void *data, size_t length, int dataSampleRate, int dataBitDepth, int dataChannels);
 
 private:
